# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: FlawFinder AI API
  description: The Ruthless Business & ML Critic - Backend API
  version: 1.0.0
  contact:
    name: FlawFinder AI Team
    email: support@flawfinder.ai
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000/api/v1
    description: Development server
  - url: https://api.flawfinder.ai/api/v1
    description: Production server
paths:
...
  /flows:
    get:
      summary: Get all flows for current user
      tags: [flows]
      security:
        - BearerAuth: []
      parameters:
        - name: skip
          in: query
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
      responses:
        '200':
          description: List of flows
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProcessFlowResponse'

    post:
      summary: Create a new process flow
      tags: [flows]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProcessFlowCreate'
      responses:
        '201':
          description: Flow created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcessFlowResponse'
  /flows/{flow_id}:
    get:
      summary: Get a specific flow
      tags: [flows]
      security:
        - BearerAuth: []
      parameters:
        - name: flow_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Flow details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcessFlowResponse'
  /flows/{flow_id}/parse:
    post:
      summary: Parse a process flow
      tags: [flows]
      security:
        - BearerAuth: []
      parameters:
        - name: flow_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Flow parsed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FlowParseResponse'
  /flows/{flow_id}/score:
    get:
      summary: Get brutality score for a flow
      tags: [flows]
      security:
        - BearerAuth: []
      parameters:
        - name: flow_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Brutality score
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BrutalityScoreResponse'
  /flows/{flow_id}/fixes:
    get:
      summary: Get role-based fixes for a flow
      tags: [flows]
      security:
        - BearerAuth: []
      parameters:
        - name: flow_id
          in: path
          required: true
          schema:
            type: integer
        - name: role
          in: query
          required: true
          schema:
            type: string
            enum: [engineer, ops, data, exec]
      responses:
        '200':
          description: Role-based fixes
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object

  # Reports endpoints
  /reports/ceo:
    post:
      summary: Generate CEO report
      tags: [reports]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CEOReportRequest'
      responses:
        '201':
          description: CEO report generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportResponse'
  /reports:
    get:
      summary: Get all reports for current user
      tags: [reports]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReportResponse'
  /reports/{report_id}:
    get:
      summary: Get a specific report
      tags: [reports]
      security:
        - BearerAuth: []
      parameters:
        - name: report_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Report details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportResponse'
  /reports/{report_id}/download:
    get:
      summary: Download report file
      tags: [reports]
      security:
        - BearerAuth: []
      parameters:
        - name: report_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Report download URL
          content:
            application/json:
              schema:
                type: object
                properties:
                  download_url:
                    type: string

  # AgentOps endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ProcessFlowCreate:
      type: object
      required:
        - name
        - raw_input
      properties:
        name:
          type: string
        description:
          type: string
        raw_input:
          type: string
        organization_id:
          type: integer
    ProcessFlowResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        version:
          type: integer
        status:
          type: string
        brutality_score:
          type: number
        user_id:
          type: integer
        organization_id:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    FlowParseResponse:
      type: object
      properties:
        flow_id:
          type: integer
        parsed_graph:
          type: object
        findings:
          type: array
          items:
            type: object
        status:
          type: string
    BrutalityScoreResponse:
      type: object
      properties:
        flow_id:
          type: integer
        score:
          type: number
        breakdown:
          type: object
        version:
          type: string

    # Report schemas
    CEOReportRequest:
      type: object
      properties:
        flow_id:
          type: integer
        include_departments:
          type: boolean
          default: true
        format:
          type: string
          enum: [pdf, json]
          default: pdf
    ReportResponse:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        report_type:
          type: string
        status:
          type: string
        content:
          type: object
        file_url:
          type: string
        metadata:
          type: object
        user_id:
          type: integer
        process_flow_id:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    # AgentOps schemas
...